# web-console/be/src/api-specifications/api-spec.yml
openapi: "3.0.3"
info:
  title: daasiot-dccs
  version: "1.2.1"

  contact:
    name: API Support
    email: a.farfaglia@sebyone.it

  license:
    name: MPL-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
servers:
  - url: http://localhost:{port}/api
    description: Development server
    variables:
      port:
        default: "3000"
        description: default port for the development server
...
components:
...
  responses:
    InternalError:
      description: Errore interno del server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    DinLocalFull:
      allOf:
        - $ref: "#/components/schemas/DinLocal"
        - type: object
          properties:
            din:
              $ref: "#/components/schemas/Din"
            links:
              type: array
              items:
                $ref: "#/components/schemas/DinLink"
            device:
              $ref: "#/components/schemas/Device"
              nullable: true
    DinLink:
      type: object
      properties:
        id:
          type: integer
          description: ID del link
        url:
          type: string
          description: dati del link, es. IP in caso di link di tipo 2
          example: "127.0.0.1:2001"
        link:
          type: integer
          description: tipo di link
          enum: [1, 2, 3, 4]
          example: 2
        din_id:
          type: integer
          description: ID del DIN al quale è associato il link
    Din:
      type: object
      properties:
        id:
          type: integer
          description: ID del DIN
        sid:
          type: string
          description: identificatore della rete daas (SID)
        din:
          type: string
          description: identificatore univoco del nodo daas all'interno della rete
        p_res:
          type: string
          description: preferenze di rete dell'utente, (riservatezza,
          example: "000"
        skey:
          type: string
          description: SKEY del DIN
          example: "9efbaeb2a94f"
    DinWithLinks:
      allOf:
        - $ref: "#/components/schemas/Din"
        - type: object
          properties:
            links:
              type: array
              items:
                $ref: "#/components/schemas/DinLink"
...
    Device:
      allOf:
        - $ref: "#/components/schemas/DeviceCreationObject"
        - type: object
          properties:
            id:
              type: integer
              description: ID del Device
...
    Error:
      type: object
      properties:
        error_name:
          type: string
          description: Nome dell'errore
          example: "TypeError"
        message:
          type: string
          description: Messaggio di errore

#region Paths
paths:
...
  /receivers:
    get:
      description: Ritorna una lista di tutti i receivers.
      tags:
        - receivers
      responses:
        "200":
          description: Lista di tutti i receivers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DinLocalFull"

        default:
          $ref: "#/components/responses/InternalError"

    post:
      description: Crea un nuovo receiver
      tags:
        - receivers
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                din:
                  $ref: "#/components/schemas/Din"
                title:
                  type: string
                  description: Nome del receiver
              required:
                - din
                - title
              example:
                din: { "sid": "100", "din": "87" }
                title: "Receiver 1"
      responses:
        "200":
          description: Receiver aggiunto con successo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DinLocalFull"
        "400":
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"
  /receivers/count:
    get:
      description: Ritorna il numero di receivers presenti.
      tags:
        - receivers
      responses:
        "200":
          description: Numero di receivers presenti.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    description: Numero di receivers presenti
        default:
          $ref: "#/components/responses/InternalError"
  /receivers/{receiverId}:
    get:
      description: Ritorna un receiver specifico.
      tags:
        - receivers
      parameters:
        - name: receiverId
          in: path
          description: ID del receiver
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receiver trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DinLocalFull"
                description: Receiver completo con i dati del DIN associato e i link
        "404":
          description: Receiver non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"

    put:
      description: Aggiorna un receiver specifico.
      tags:
        - receivers
      parameters:
        - name: receiverId
          in: path
          description: ID del receiver
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                din:
                  $ref: "#/components/schemas/Din"
                title:
                  type: string
                  description: Nome del receiver
                enable:
                  type: boolean
                  description: Flag per abilitare/disabilitare il receiver
                acpt_all:
                  type: boolean
                  description: Flag per accettare messaggi da nodi non mappati
              required:
                # TODO: rendere il din non obbligatorio
                - din
                - title
              example:
                din: { "sid": "100", "din": "87" }
                title: "Receiver 1"
      responses:
        "200":
          description: Receiver aggiornato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio di successo
        "400":
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Receiver o Din non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"

    delete:
      description: Elimina un receiver specifico.
      tags:
        - receivers
      parameters:
        - name: receiverId
          in: path
          description: ID del receiver
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receiver eliminato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio di successo
        "404":
          description: Receiver non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"

  #region Receivers links
  /receivers/{receiverId}/links:
    get:
      description: Ritorna i link associati al receiver.
      tags:
        - receivers
      parameters:
        - name: receiverId
          in: path
          description: ID del receiver
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista di tutti i link associati al receiver.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DinLink"
        "404":
          description: Receiver non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"

    post:
      description: Aggiunge un link al receiver.
      tags:
        - receivers
      parameters:
        - name: receiverId
          in: path
          description: ID del receiver
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: URL del link
                link:
                  type: integer
                  description: Tipo di link
                  enum: [1, 2, 3, 4]
              required:
                - url
                - link
              example:
                link: 2
                url: "151.81.1.4:2001"
      responses:
        "200":
          description: Link aggiunto con successo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DinLink"
        "400":
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Receiver non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"
...
  /dins:
    get:
      description: Ritorna una lista di tutti i DINs.
      tags:
        - receivers
        - remotes
      responses:
        "200":
          description: Lista di tutti i DINs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Din"
        default:
          $ref: "#/components/responses/InternalError"
  /remotes:
    get:
      description: Ritorna una lista di tutti i nodi remoti, ovvero i DINs non associati ad un receiver.
      tags:
        - remotes
      responses:
        "200":
          description: Lista di tutti i nodi remoti.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DinWithLinks"

        default:
          $ref: "#/components/responses/InternalError"
  /remotes/count:
    get:
      description: Ritorna il numero di nodi remoti presenti.
      tags:
        - remotes
      responses:
        "200":
          description: Numero di nodi remoti presenti.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    description: Numero di nodi remoti presenti
        default:
          $ref: "#/components/responses/InternalError"

  #region Receivers remotes
...
  /receivers/{receiverId}/remotes/{id}:
    get:
      description: Ritorna un nodo remoto specifico associato al receiver.
      tags:
        - receivers
        - remotes
      parameters:
        - name: receiverId
          in: path
          description: ID del receiver
          required: true
          schema:
            type: integer
        - name: id
          in: path
          description: ID del nodo remoto
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Nodo remoto trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DinMappingFull"
        "404":
          description: Receiver o nodo remoto mappato a receiver non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"

    delete:
      description: Elimina l'associazione tra il nodo remoto e il receiver, se il nodo remoto non ha altre associazioni, viene cancellato.
      tags:
        - receivers
        - remotes
      parameters:
        - name: receiverId
          in: path
          description: ID del receiver
          required: true
          schema:
            type: integer
        - name: id
          in: path
          description: ID del nodo remoto
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Nodo remoto eliminato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio di successo
        "404":
          description: Receiver o nodo remoto non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"

  #region Devices
...